Create Install_AnyConnect.vbs
An example of a bit more mature (version 3) of VBScript which performs an application installation, with checking of existence of pre-requisite, another, application, and with caching installation file(s) locally (which is very helpful for VPN users, who might get disconnected during the installation).

This example installs Cisco AnyConnect from DFS. But the pre-reqs are there, e.g. you can use APP_PREREQ to copy the configuration file for Umbrella.